Update Treasury-LC.lua
As cute as the name-hashing function was, it was too unreliable to keep. I still can't see a single reason to put a prepare() statement in the scheduling function since the function it's calling isn't pre-prepared elsewhere, but I reverted it to the original form anyway just in case. Did not restore the constantly-breaking varargs or incomprehensible (and also constantly-breaking) lot/pass/drop metagarbage though.